Move jSerialPort implementation to its own package
