+++ java/test/jmri/ArchitectureTest.java
@@ -311,8 +311,8 @@ public class ArchitectureTest {
         .that()
 
         // all the standard serial access should be confined to here:
-        .doNotHaveFullyQualifiedName("jmri.jmrix.SerialPort").and()
-        .doNotHaveFullyQualifiedName("jmri.jmrix.SerialPortEvent")
+        .doNotHaveFullyQualifiedName("jmri.jmrix.jserialcomm.JSerialPort").and()
+        .doNotHaveFullyQualifiedName("jmri.jmrix.jserialcomm.JSerialPortEvent")
 
         .should().accessClassesThat().resideInAPackage("com.fazecast.jSerialComm..");
